+++ 3d/background.js
@@ -1,14 +1,8 @@
+import { cubeVertices } from "./cube.js";
+
 export default class Background {
 
-    static cubeVertices = new Float32Array([
-        // 36 vertices (12 triangles) for a cube, positions only
-        -1, -1, -1, 1, -1, -1, 1, 1, -1, -1, -1, -1, 1, 1, -1, -1, 1, -1, // back
-        -1, -1, 1, 1, -1, 1, 1, 1, 1, -1, -1, 1, 1, 1, 1, -1, 1, 1, // front
-        -1, 1, -1, 1, 1, -1, 1, 1, 1, -1, 1, -1, 1, 1, 1, -1, 1, 1, // top
-        -1, -1, -1, 1, -1, -1, 1, -1, 1, -1, -1, -1, 1, -1, 1, -1, -1, 1, // bottom
-        1, -1, -1, 1, 1, -1, 1, 1, 1, 1, -1, -1, 1, 1, 1, 1, -1, 1, // right
-        -1, -1, -1, -1, 1, -1, -1, 1, 1, -1, -1, -1, -1, 1, 1, -1, -1, 1  // left
-    ]);
+    
 
     static async fromPaths(gpu, { image, shader, projectionMatrixBuffer }) {
         const texture = await gpu.createCubeTexture([image, image, image, image, image, image], "rgba8unorm");
@@ -20,17 +14,11 @@ export default class Background {
     constructor(gpu, { texture, module, projectionMatrixBuffer }) {
         this.gpu = gpu;
         this.texture = texture;
-        // this.buffer = gpu.createUniformBuffer(144);
         this.projectionMatrixBuffer = projectionMatrixBuffer;
         this.sampler = gpu.createSampler();
         this.pipeline = gpu.createRenderPipeline(module, "vsMain", module, "fsMain");
-        this.vertexBuffer = gpu.device.createBuffer({
-            size: Background.cubeVertices.byteLength,
-            usage: GPUBufferUsage.VERTEX,
-            mappedAtCreation: true
-        });
-        new Float32Array(this.vertexBuffer.getMappedRange()).set(Background.cubeVertices);        
-        this.vertexBuffer.unmap();
+        this.vertexBuffer = gpu.createVertexBuffer(cubeVertices.byteLength)
+        this.gpu.device.queue.writeBuffer(this.vertexBuffer, 0, cubeVertices);
 
         this.bindGroup = gpu.createBindGroup({
             layout: this.pipeline.getBindGroupLayout(0),
@@ -50,7 +38,7 @@ export default class Background {
         pass.setPipeline(this.pipeline);
         pass.setBindGroup(0, this.bindGroup);
         pass.setVertexBuffer(0, this.vertexBuffer);
-        pass.draw(36, 1, 0, 0); // draw the cube
+        pass.draw(36, 1, 0, 0);
     }
